got rid of WApplication.
This is in preparation for WSession to become the new
Application.  I used to have a split but it wasn't really
right, so merged most of both of them into WSession.  Well the problem
is that WSession really is a session for executing the bbs, not
really a "user session", which I think is really what we need This way
we can cleanup after a user is finished (i.e. from after GotCaller() until
the end of the loop.

move wfc into it's own class

{G,S}etWfcStatus -> at_wfc

cleanup

+++ bbs/dropfile.cpp
@@ -33,6 +33,7 @@
 #include "sdk/status.h"
 #include "core/strings.h"
 #include "core/textfile.h"
+#include "core/version.h"
 
 using std::string;
 using namespace wwiv::sdk;
@@ -502,9 +503,9 @@ const string create_chain_file() {
 
   create_drop_files();
   auto start_duration = duration_since_midnight(session()->system_logon_time());
-  int start_second = std::chrono::duration_cast<std::chrono::seconds>(start_duration).count();
+  int start_second = static_cast<int>(std::chrono::duration_cast<std::chrono::seconds>(start_duration).count());
   auto used_duration = std::chrono::system_clock::now() - session()->system_logon_time();
-  int seconds_used = std::chrono::duration_cast<std::chrono::seconds>(used_duration).count();
+  int seconds_used = static_cast<int>(std::chrono::duration_cast<std::chrono::seconds>(used_duration).count());
 
   const string fileName = create_filename(CHAINFILE_CHAIN);
   File::Remove(fileName);
